+++ spec/speclj/tags_spec.clj
@@ -73,4 +73,6 @@
 
   )
 
+blah
+
 (run-specs)