+++ CHANGES.md
@@ -7,8 +7,10 @@
 * renamed specdoc to documentation
 * output improved
 * with-all spec component
+* tag examples. declared in describe blocks. filtered on run.
+* exclude hidden files in runs (https://github.com/slagyr/speclj/pull/4)
 
-# 1.3.0
+ # 1.3.0
 
 * using raw Runtime.exec in 'lein spec' command to avoid ant's output mangling and lag.
 * should-be-same or some assertion that checks object reference equality